+++ linphone-app/assets/languages/ru.ts
@@ -1018,7 +1018,7 @@
     </message>
     <message>
         <source>conversationMenuDelete</source>
-        <extracomment>&apos;Delete&apos; : Item menu to delete the chat</extracomment>
+        <extracomment>&apos;Delete history&apos; : Item menu to delete the chat&apos;s history</extracomment>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -3401,6 +3401,19 @@
         <source>timelineFilterConferences</source>
         <extracomment>&apos;Conferences&apos; : Filter item. Selecting it will show all conferences.</extracomment>
         <translation>Конференции</translation>
+    </message>
+</context>
+<context>
+    <name>TimelineItem</name>
+    <message>
+        <source>deleteTimeline</source>
+        <extracomment>&apos;Are you sure you want to delete and leave this timeline?&apos;</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>deleteTimelineTooltip</source>
+        <extracomment>&apos;After confirmation, it will erase all history, leave the chat room if it is a group chat and delete it in database.&apos;</extracomment>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>